+++ dotnet/samples/GettingStartedWithAgents/OpenAIAssistant/Step07_Assistant_Declarative.cs
@@ -39,7 +39,7 @@ public class Step07_Assistant_Declarative : BaseAssistantTest
             """;
         OpenAIAssistantAgentFactory factory = new();
 
-        var agent = await factory.CreateAgentFromYamlAsync(text, new() { Configuration = TestConfiguration.ConfigurationRoot });
+        var agent = await factory.CreateAgentFromYamlAsync(text, configuration: TestConfiguration.ConfigurationRoot);
 
         await InvokeAgentAsync(agent!, "Could you please create a bar chart for the operating profit using the following data and provide the file to me? Company A: $1.2 million, Company B: $2.5 million, Company C: $3.0 million, Company D: $1.8 million");
     }
@@ -65,7 +65,7 @@ public class Step07_Assistant_Declarative : BaseAssistantTest
         builder.Services.AddSingleton<TokenCredential>(new AzureCliCredential());
         var kernel = builder.Build();
 
-        var agent = await factory.CreateAgentFromYamlAsync(text, new() { Kernel = kernel, Configuration = TestConfiguration.ConfigurationRoot });
+        var agent = await factory.CreateAgentFromYamlAsync(text, new() { Kernel = kernel }, TestConfiguration.ConfigurationRoot);
 
         await InvokeAgentAsync(agent!, "Could you please create a bar chart for the operating profit using the following data and provide the file to me? Company A: $1.2 million, Company B: $2.5 million, Company C: $3.0 million, Company D: $1.8 million");
     }
@@ -84,7 +84,7 @@ public class Step07_Assistant_Declarative : BaseAssistantTest
             """;
         OpenAIAssistantAgentFactory factory = new();
 
-        var agent = await factory.CreateAgentFromYamlAsync(text, new() { Kernel = this._kernel, Configuration = TestConfiguration.ConfigurationRoot });
+        var agent = await factory.CreateAgentFromYamlAsync(text, new() { Kernel = this._kernel }, configuration: TestConfiguration.ConfigurationRoot);
 
         await InvokeAgentAsync(agent!, "Cats and Dogs");
     }
@@ -117,7 +117,7 @@ public class Step07_Assistant_Declarative : BaseAssistantTest
         OpenAIAssistantAgentFactory factory = new();
         var promptTemplateFactory = new KernelPromptTemplateFactory();
 
-        var agent = await factory.CreateAgentFromYamlAsync(text, new() { Kernel = this._kernel, Configuration = TestConfiguration.ConfigurationRoot, PromptTemplateFactory = promptTemplateFactory });
+        var agent = await factory.CreateAgentFromYamlAsync(text, new() { Kernel = this._kernel, PromptTemplateFactory = promptTemplateFactory }, TestConfiguration.ConfigurationRoot);
         Assert.NotNull(agent);
 
         var options = new AgentInvokeOptions()
