+++ dotnet/src/Agents/Abstractions/Definition/AgentCreationOptions.cs
@@ -1,6 +1,5 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 using System.Diagnostics.CodeAnalysis;
-using Microsoft.Extensions.Configuration;
 
 namespace Microsoft.SemanticKernel.Agents;
 
@@ -24,9 +23,4 @@ public class AgentCreationOptions
     /// Gets or sets the <see cref="IPromptTemplateFactory"/>, a factory for prompt templates for one or more prompt template formats.
     /// </summary>
     public IPromptTemplateFactory? PromptTemplateFactory { get; init; } = null;
-
-    /// <summary>
-    /// Gets or sets the <see cref="IConfiguration"/>, a set of key/value agent configuration properties.
-    /// </summary>
-    public IConfiguration? Configuration { get; init; } = null;
 }
