+++ dotnet/src/Agents/Yaml/Agents.Yaml.csproj
@@ -28,6 +28,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" />
     <PackageReference Include="YamlDotNet" />
   </ItemGroup>
