+++ dotnet/src/Agents/Yaml/Extensions/YamlKernelAgentFactoryExtensions.cs
@@ -2,24 +2,26 @@
 
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Configuration;
 
 namespace Microsoft.SemanticKernel.Agents;
 
 /// <summary>
 /// Extension methods for <see cref="KernelAgentFactory"/> to create agents from YAML.
 /// </summary>
-public static class KernelAgentFactoryYamlExtensions
+public static class YamlKernelAgentFactoryExtensions
 {
     /// <summary>
     /// Create a <see cref="KernelAgent"/> from the given YAML text.
     /// </summary>
-    /// <param name="kernelAgentFactory">Kernel agent factory which will be used to create the agent</param>
+    /// <param name="kernelAgentFactory">Kernel agent factory which will be used to create the agent.</param>
     /// <param name="text">Text string containing the YAML representation of a kernel agent.</param>
-    /// <param name="options">Optional <see cref="AgentCreationOptions"/> instance</param>
+    /// <param name="options">Optional <see cref="AgentCreationOptions"/> instance.</param>
+    /// <param name="configuration">Optional <see cref="IConfiguration"/> instance.</param>
     /// <param name="cancellationToken">Optional cancellation token</param>
-    public static async Task<KernelAgent?> CreateAgentFromYamlAsync(this KernelAgentFactory kernelAgentFactory, string text, AgentCreationOptions? options = null, CancellationToken cancellationToken = default)
+    public static async Task<KernelAgent?> CreateAgentFromYamlAsync(this KernelAgentFactory kernelAgentFactory, string text, AgentCreationOptions? options = null, IConfiguration? configuration = null, CancellationToken cancellationToken = default)
     {
-        var agentDefinition = AgentDefinitionYaml.FromYaml(text, options?.Configuration);
+        var agentDefinition = AgentDefinitionYaml.FromYaml(text, configuration);
         agentDefinition.Type ??= (kernelAgentFactory.Types.Count > 0 ? kernelAgentFactory.Types[0] : null);
 
         return await kernelAgentFactory.CreateAsync(
